requirements.txt
customtkinter==5.2.2
matplotlib==3.10.8
pandas==2.3.3
Requests==2.32.5
websocket_client==1.9.0
